[FEAT]: Add Sliding window codding pattern question.
Substring with Concatenation of All Words.
You are given a string, s, and a list of words, words, that are all of the same
length. Find all starting indices of substring(s) in s that is a concatenation of
each word in words exactly once and without any intervening characters.